Refactor CUDA array and memory management
- Moved array and texture management functions from detail namespace to cuda namespace for better accessibility.
- Updated array and texture creation and deletion functions to use new naming conventions.
- Changed data types from u32 to size_t for better compatibility with larger data sizes.
- Improved memory management in FFT functions by updating parameter types and ensuring consistency across function signatures.
- Refactored NdView and NdArray classes to use size_t instead of u32 for dimensions and sizes, enhancing flexibility.
- Cleaned up CUDA stream and event management functions, removing unnecessary detail namespace.
- Added new math library for FFT operations, ensuring proper integration with CUDA.
- Updated FDK reconstruction algorithms to utilize new CUDA types and functions, improving performance and readability.

+++ src/nct/cuda/fft.h
@@ -6,9 +6,9 @@ namespace nct::cuda {
 
 using fft_plan_t = int;
 
-auto fft_plan_c2c(const u32 (&dim)[1], u32 batch) -> fft_plan_t;
-auto fft_plan_r2c(const u32 (&dim)[1], u32 batch) -> fft_plan_t;
-auto fft_plan_c2r(const u32 (&dim)[1], u32 batch) -> fft_plan_t;
+auto fft_plan_c2c(const size_t (&dim)[1], size_t batch) -> fft_plan_t;
+auto fft_plan_r2c(const size_t (&dim)[1], size_t batch) -> fft_plan_t;
+auto fft_plan_c2r(const size_t (&dim)[1], size_t batch) -> fft_plan_t;
 
 void fft_destroy(fft_plan_t plan);
